fix warnings for arm32

+++ src/seq_code.c
@@ -40,13 +40,13 @@ struct seq_code const* seq_code_create(struct imm_abc const* abc, uint8_t min_se
     }
 
     if (max_seq > 1) {
-        for (unsigned len = max_seq - 2; 1 <= len + 1; --len)
+        for (unsigned len = (unsigned) (max_seq - 2); 1 <= len + 1; --len)
             seq_code->stride[len] = seq_code->stride[len + 1] * imm_abc_length(abc);
     }
 
     seq_code->offset = malloc(sizeof(*seq_code->offset) * offset_size(seq_code));
     seq_code->offset[0] = 0;
-    for (unsigned len = min_seq + 1; len <= max_seq; ++len) {
+    for (unsigned len = (unsigned) (min_seq + 1); len <= max_seq; ++len) {
 
         unsigned i = len - min_seq;
         seq_code->offset[i] = seq_code->offset[i - 1] + seq_code->stride[max_seq - (len - 1) - 1];
@@ -76,7 +76,7 @@ struct eseq* seq_code_create_eseq(struct seq_code const* seq_code)
 {
     struct eseq* eseq = malloc(sizeof(*eseq));
     eseq->seq_code = seq_code;
-    eseq->code = matrixu_create(1, seq_code->max_seq - seq_code->min_seq + 1);
+    eseq->code = matrixu_create(1, (unsigned) (seq_code->max_seq - seq_code->min_seq + 1));
     return eseq;
 }
 
@@ -190,5 +190,5 @@ err:
 
 static inline uint8_t offset_size(struct seq_code const* seq_code)
 {
-    return seq_code->max_seq - seq_code->min_seq + 1;
+    return (uint8_t) (seq_code->max_seq - seq_code->min_seq + 1);
 }
